+++ tests/cpp_caching/op_user_lib.h
@@ -1,1 +1,1 @@
-float use_op();
+int use_op();
